+++ src/or/rendmid.c
@@ -13,13 +13,13 @@
  * setting the circuit's purpose and service pk digest.
  */
 int
-rend_mid_establish_intro(circuit_t *circ, const char *request, int request_len)
+rend_mid_establish_intro(circuit_t *circ, const char *request, size_t request_len)
 {
   crypto_pk_env_t *pk = NULL;
   char buf[DIGEST_LEN+9];
   char expected_digest[DIGEST_LEN];
   char pk_digest[DIGEST_LEN];
-  int asn1len;
+  size_t asn1len;
   circuit_t *c;
   char serviceid[REND_SERVICE_ID_LEN+1];
 
@@ -110,7 +110,7 @@ rend_mid_establish_intro(circuit_t *circ, const char *request, int request_len)
  * INTRODUCE2 cell.
  */
 int
-rend_mid_introduce(circuit_t *circ, const char *request, int request_len)
+rend_mid_introduce(circuit_t *circ, const char *request, size_t request_len)
 {
   circuit_t *intro_circ;
   char serviceid[REND_SERVICE_ID_LEN+1];
@@ -177,7 +177,7 @@ rend_mid_introduce(circuit_t *circ, const char *request, int request_len)
  * rendezvous cookie.
  */
 int
-rend_mid_establish_rendezvous(circuit_t *circ, const char *request, int request_len)
+rend_mid_establish_rendezvous(circuit_t *circ, const char *request, size_t request_len)
 {
   char hexid[9];
 
@@ -224,7 +224,7 @@ rend_mid_establish_rendezvous(circuit_t *circ, const char *request, int request_
  * connecting the two circuits.
  */
 int
-rend_mid_rendezvous(circuit_t *circ, const char *request, int request_len)
+rend_mid_rendezvous(circuit_t *circ, const char *request, size_t request_len)
 {
   circuit_t *rend_circ;
   char hexid[9];
